# Added an optimization for search results