--- a/scripts/artifacts/sms.py
+++ b/scripts/artifacts/sms.py
@@ -59,7 +59,7 @@ def get_sms(files_found, report_folder, seeker):
         def copyAttachments(rec):
             pathToAttachment = None
             if rec["FILENAME"]:
-                attachment = seeker.search('**'+rec["FILENAME"].replace('~',''))
+                attachment = seeker.search('**'+rec["FILENAME"].replace('~',''), return_on_first_hit=True)
                 pathToAttachment = os.path.join((os.path.basename(os.path.abspath(report_folder))), os.path.basename(rec["FILENAME"]))
                 if is_platform_windows():
                     invalid = '<>:"/\|?*'