--- a/scripts/ilapfuncs.py
+++ b/scripts/ilapfuncs.py
@@ -288,7 +288,7 @@ searching for thumbnails, copy it to report folder and return tag  to insert in 
 '''
 def generate_thumbnail(imDirectory, imFilename, seeker, report_folder):
     thumb = thumbnail_root+imDirectory+'/'+imFilename+'/'
-    thumblist = seeker.search(thumb+'**.JPG')
+    thumblist = seeker.search(thumb+'**.JPG', return_on_first_hit=True)
     thumbname = imDirectory.replace('/','_')+'_'+imFilename+'.JPG'
     pathToThumb = os.path.join(os.path.basename(os.path.abspath(report_folder)), thumbname)
     htmlThumbTag = '<img src="{0}"></img>'.format(pathToThumb)
@@ -297,7 +297,7 @@ def generate_thumbnail(imDirectory, imFilename, seeker, report_folder):
     else:
         #recreate thumbnail from image
         #TODO: handle videos and HEIC
-        files = seeker.search(media_root+imDirectory+'/'+imFilename)
+        files = seeker.search(media_root+imDirectory+'/'+imFilename, return_on_first_hit=True)
         if files:
             try:
                 im = Image.open(files[0])