Now there is an option to terminate the file search early, if you only need the first search result. The seeker.search has this option added, function is now: `def search(self, filepattern, return_on_first_hit=False)` 

For at least a couple of plugins, this results in much faster processing.

Using filesystem extraction type on the 'Ruth' ctf ios image.

**PhotosMetadata plugin**
563.5 sec - Before 
312.4 sec - Now

**SMS plugin**
25.51 sec - Before
13.59 sec - Now

I've also sorted the modules under `ilap_artifacts.py` for easy reading/editing.